feeder-research/requirements.txt
pyserial
pandas
numpy
matplotlib
plotly
jupyter
openpyxl
ipykernel
